+++ src/providers/linkProvider.ts
@@ -19,7 +19,7 @@ export default class LinkProvider implements vsDocumentLinkProvider {
 
     async provideDocumentLinks(doc: TextDocument): Promise<DocumentLink[]> {
         let range = window.activeTextEditor.visibleRanges[0]
-        let reg = new RegExp(`(?<=(${this.regex}))\\(['"](.*?)['"]`, 'g')
+        let reg = new RegExp(`(?<=(${this.regex})\\()['"](.*?)['"]`, 'g')
         let documentLinks = []
 
         for (let i = range.start.line; i <= range.end.line; i++) {
@@ -27,17 +27,17 @@ export default class LinkProvider implements vsDocumentLinkProvider {
             let txt = line.text
             let result = txt.match(reg)
 
-            if (result != null) {
+            if (result) {
                 for (let found of result) {
-                    let files = await util.getFilePaths(found, doc)
+                    if (!found.includes('*')) {
+                        let file = await util.getFilePath(found, doc)
 
-                    if (files.length) {
-                        let start = new Position(line.lineNumber, txt.indexOf(found))
-                        let end = start.translate(0, found.length)
+                        if (file) {
+                            let start = new Position(line.lineNumber, txt.indexOf(found))
+                            let end = start.translate(0, found.length)
 
-                        for (const file of files) {
                             let documentlink = new DocumentLink(new Range(start, end), file.fileUri)
-                            documentlink.tooltip = file.showPath
+                            documentlink.tooltip = file.tooltip
                             documentLinks.push(documentlink)
                         }
                     }