+++ CHANGELOG.md
@@ -8,3 +8,8 @@
 
 - add config for methods search
 - show underline under the key only instead of the whole line
+
+## 0.0.5
+
+- create view if not found
+- fix loss search, now it will get exact match